# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Karm Baba B2B Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
  /users/me/onboarding:
    post:
      operationId: completeUserOnboarding
      tags: [users]
      summary: Choose buyer or seller role (explicit onboarding)
      description: >
        Sets the authenticated user's marketplace role. Role is stored in Postgres
        and is not overwritten by opaque Clerk publicMetadata on sync (except admin elevation).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
        "403":
          description: Admins cannot change role via onboarding
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, onboardingCompleted, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
        company:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
    OnboardingInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [buyer, seller]
          description: Marketplace role chosen by the user in onboarding UI.
        company:
          type: string
          description: Optional company / trading name.
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [buyer, seller]
        company:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...